requirements.txt
Flask>=2.0
python-dotenv>=0.15
Flask-SQLAlchemy>=2.5
Flask-Login>=0.5
google-generativeai>=0.3
reportlab>=3.6
PyPDF2>=3.0
Werkzeug>=2.0
